+++ R/startup.R
@@ -3,7 +3,8 @@ printRIstartupInfo <- function()
   version <- packageVersion('easyRasch')
   hello <- paste("This is easyRasch ",version,
                  ". Check for updates at <https://pgmj.github.io/easyRasch/news/index.html> \n",
-                 "Remember to use `set.seed()` at the start of your code/document."
+                 "Remember to use `set.seed()` at the start of your code/document.\n",
+                 "MacOS users with Tahoe 26.1 or later need to add/run `options(rgl.useNULL = TRUE)` BEFORE loading easyRasch"
                  ,sep="")
   packageStartupMessage(hello)
 }
@@ -11,3 +12,7 @@ printRIstartupInfo <- function()
 .onAttach <- function(...) {
   printRIstartupInfo()
 }
+
+# .onLoad <- function(...) {
+#   options(rgl.useNULL = TRUE)
+# }